Password Length
The default password length for devise is 8 characters be the flash messages are not displaying in the login layout. This adds a message to the registration field stating the password length. Not closing this as the rails flash message isn't being displayed, but this should help until the issue is fully resolved.

references #5